refactor: 556 - update structure mentorship routes
